# docs/api/catalog-service.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "description": "Kitap kataloğu: herkese açık okuma, admin yönetimi ve servisler arası stok rezervasyonu.\n\n**Erişim:**\n- Herkese açık (`GET /api/books/**`, `GET /api/categories/**`): kimlik gerekmez; `Authorization` başlığı gönderilse de yok sayılır.\n- Admin (`/api/admin/**`): user-service'ten alınan JWT, `Authorization: Bearer <accessToken>`, `role` = `ADMIN` (`bearerAuth`).\n- Internal (`/internal/**`): yalnızca servisler arası, `X-Internal-Api-Key` başlığı (`internalApiKey`). Kullanıcı JWT'si (ADMIN dahil) bu uçlarda geçersizdir. Ayrıntılar: `docs/api/catalog-internal-stock.md`.\n\n**Hatalar:** Tüm hatalar RFC 9457 ProblemDetail olarak `application/problem+json` ile döner (`Problem` şeması). İstemci hatayı `code` alanına göre ayırt eder; doğrulama hatalarında `errors` dizisi alan bazında mesaj içerir. Stok rezervasyonu çakışmalarında ek olarak `bookIds` döner (`StockUnavailableProblem`).\n",
    "title": "Kitap Sepeti Catalog API",
    "version": "0.0.1-SNAPSHOT"
  },
  "servers": [
    {
      "url": "/",
      "description": "Dokümanın sunulduğu sunucu"
    }
  ],
...
  "paths": {
...
    "/api/admin/books/{id}": {
      "delete": {
        "description": "Arşivlemeyle aynı işlem; kayıt fiziksel olarak silinmez.",
        "operationId": "deleteBook",
        "parameters": [
          {
            "description": "Kitap id'si",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Arşivlendi."
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`VALIDATION_FAILED` (alan hataları `errors` dizisinde) veya `MALFORMED_REQUEST` (okunamayan JSON ya da bozuk path değişkeni)."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`UNAUTHORIZED`: token yok, geçersiz ya da süresi dolmuş.",
            "headers": {
              "WWW-Authenticate": {
                "description": "Token yoksa `Bearer`, geçersizse `Bearer error=\"invalid_token\"`.",
                "schema": {
                  "type": "string",
                  "example": "Bearer"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`FORBIDDEN`: token geçerli ama rol `ADMIN` değil."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`RESOURCE_NOT_FOUND`: kayıt yok."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`INTERNAL_ERROR`: beklenmeyen hata; ayrıntı yanıtta yer almaz."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`AUTHENTICATION_UNAVAILABLE`: token doğrulanamadı çünkü user-service JWKS ucuna ulaşılamıyor; daha sonra tekrar deneyin."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Sil (= arşivle)",
        "tags": [
          "Admin – Books"
        ]
      },
      "get": {
        "operationId": "getAdminBook",
        "parameters": [
          {
            "description": "Kitap id'si",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminBookResponse"
                }
              }
            },
            "description": "Kitap."
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`VALIDATION_FAILED` (alan hataları `errors` dizisinde) veya `MALFORMED_REQUEST` (okunamayan JSON ya da bozuk path değişkeni)."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`UNAUTHORIZED`: token yok, geçersiz ya da süresi dolmuş.",
            "headers": {
              "WWW-Authenticate": {
                "description": "Token yoksa `Bearer`, geçersizse `Bearer error=\"invalid_token\"`.",
                "schema": {
                  "type": "string",
                  "example": "Bearer"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`FORBIDDEN`: token geçerli ama rol `ADMIN` değil."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`RESOURCE_NOT_FOUND`: kayıt yok."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`INTERNAL_ERROR`: beklenmeyen hata; ayrıntı yanıtta yer almaz."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`AUTHENTICATION_UNAVAILABLE`: token doğrulanamadı çünkü user-service JWKS ucuna ulaşılamıyor; daha sonra tekrar deneyin."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Kitap getir (tüm durumlar)",
        "tags": [
          "Admin – Books"
        ]
      },
      "patch": {
        "description": "Kısmi güncelleme; `version` zorunlu ve kaydın güncel versiyonu olmalı. Yayındaki kitapta `BookUpserted` olayı yazılır.",
        "operationId": "updateBook",
        "parameters": [
          {
            "description": "Kitap id'si",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBookRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminBookResponse"
                }
              }
            },
            "description": "Güncellenmiş kitap."
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`VALIDATION_FAILED` (alan hataları `errors` dizisinde) veya `MALFORMED_REQUEST` (okunamayan JSON ya da bozuk path değişkeni)."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`UNAUTHORIZED`: token yok, geçersiz ya da süresi dolmuş.",
            "headers": {
              "WWW-Authenticate": {
                "description": "Token yoksa `Bearer`, geçersizse `Bearer error=\"invalid_token\"`.",
                "schema": {
                  "type": "string",
                  "example": "Bearer"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`FORBIDDEN`: token geçerli ama rol `ADMIN` değil."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`RESOURCE_NOT_FOUND`: kayıt yok."
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`CONCURRENT_MODIFICATION`: `version` güncel değil (hiçbir şey yazılmadı; kaydı yeniden okuyun) veya `ISBN_ALREADY_EXISTS`."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`INTERNAL_ERROR`: beklenmeyen hata; ayrıntı yanıtta yer almaz."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`AUTHENTICATION_UNAVAILABLE`: token doğrulanamadı çünkü user-service JWKS ucuna ulaşılamıyor; daha sonra tekrar deneyin."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Kitap güncelle",
        "tags": [
          "Admin – Books"
        ]
      }
    },
...
    "/api/books": {
      "get": {
        "description": "Yalnızca yayındaki kitaplar. `categoryId` alt kategorilerdeki kitapları da kapsar. `minPrice` > `maxPrice` ise 400.",
        "operationId": "searchBooks",
        "parameters": [
          {
            "description": "Kategori; alt kategorilerdeki kitaplar da dahil",
            "in": "query",
            "name": "categoryId",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Kategori; alt kategorilerdeki kitaplar da dahil"
            }
          },
          {
            "in": "query",
            "name": "publisherId",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "authorId",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "minPrice",
            "required": false,
            "schema": {
              "type": "number",
              "minimum": 0.0
            }
          },
          {
            "description": "`minPrice` ile birlikte verilirse ondan küçük olamaz",
            "in": "query",
            "name": "maxPrice",
            "required": false,
            "schema": {
              "type": "number",
              "description": "`minPrice` ile birlikte verilirse ondan küçük olamaz",
              "minimum": 0.0
            }
          },
          {
            "description": "Büyük/küçük harf duyarsız",
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "type": "string",
              "default": "newest",
              "description": "Büyük/küçük harf duyarsız",
              "enum": [
                "newest",
                "price_asc",
                "price_desc",
                "title_asc"
              ]
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0,
              "minimum": 0
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20,
              "maximum": 50,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseBookSummaryResponse"
                }
              }
            },
            "description": "Sayfa (son sayfadan sonrası boş `items` döner)."
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`VALIDATION_FAILED` (alan hataları `errors` dizisinde) veya `MALFORMED_REQUEST` (okunamayan JSON ya da bozuk path değişkeni)."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`INTERNAL_ERROR`: beklenmeyen hata; ayrıntı yanıtta yer almaz."
          }
        },
        "security": [],
        "summary": "Kitapları listele / ara",
        "tags": [
          "Books"
        ]
      }
    },
    "/api/books/{id}": {
      "get": {
        "operationId": "getBook",
        "parameters": [
          {
            "description": "Kitap id'si",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookDetailResponse"
                }
              }
            },
            "description": "Kitap."
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`VALIDATION_FAILED` (alan hataları `errors` dizisinde) veya `MALFORMED_REQUEST` (okunamayan JSON ya da bozuk path değişkeni)."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`RESOURCE_NOT_FOUND`: kitap yok ya da yayında değil (taslak/arşiv)."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`INTERNAL_ERROR`: beklenmeyen hata; ayrıntı yanıtta yer almaz."
          }
        },
        "security": [],
        "summary": "Kitap detayı",
        "tags": [
          "Books"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdminBookResponse": {
        "type": "object",
        "properties": {
          "authors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuthorRef"
            }
          },
          "availableQuantity": {
            "type": "integer",
            "format": "int32"
          },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategoryRef"
            }
          },
          "coverUrl": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "currency": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "isbn": {
            "type": "string"
          },
          "pageCount": {
            "type": "integer",
            "format": "int32"
          },
          "priceAmount": {
            "type": "number"
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time"
          },
          "publisher": {
            "$ref": "#/components/schemas/PublisherRef"
          },
          "reservedQuantity": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published",
              "archived"
            ]
          },
          "stockQuantity": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "AuthorRef": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        }
      },
...
      "BookDetailResponse": {
        "type": "object",
        "properties": {
          "authors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuthorRef"
            }
          },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategoryRef"
            }
          },
          "coverUrl": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "inStock": {
            "type": "boolean"
          },
          "isbn": {
            "type": "string"
          },
          "pageCount": {
            "type": "integer",
            "format": "int32"
          },
          "priceAmount": {
            "type": "number"
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time"
          },
          "publisher": {
            "$ref": "#/components/schemas/PublisherRef"
          },
          "title": {
            "type": "string"
          }
        }
      },
      "BookSummaryResponse": {
        "type": "object",
        "properties": {
          "authors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuthorRef"
            }
          },
          "coverUrl": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "inStock": {
            "type": "boolean"
          },
          "priceAmount": {
            "type": "number"
          },
          "publisher": {
            "$ref": "#/components/schemas/PublisherRef"
          },
          "title": {
            "type": "string"
          }
        }
      },
      "CategoryRef": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        }
      },
...
      "FieldError": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "description": "Hatalı alanın adı.",
            "example": "slug"
          },
          "message": {
            "type": "string",
            "description": "Doğrulama mesajı.",
            "example": "must not be blank"
          }
        },
        "required": [
          "field",
          "message"
        ]
      },
...
      "PageResponseBookSummaryResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BookSummaryResponse"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "Problem": {
        "type": "object",
        "description": "RFC 9457 Problem Details. İstemci hatayı `code` alanına göre ayırt etmelidir; `title`/`detail` insan içindir ve değişebilir.",
        "properties": {
          "code": {
            "type": "string",
            "description": "Makine tarafından okunacak hata kodu.",
            "enum": [
              "VALIDATION_FAILED",
              "MALFORMED_REQUEST",
              "UNAUTHORIZED",
              "FORBIDDEN",
              "NOT_FOUND",
              "RESOURCE_NOT_FOUND",
              "METHOD_NOT_ALLOWED",
              "NOT_ACCEPTABLE",
              "UNSUPPORTED_MEDIA_TYPE",
              "CONFLICT",
              "SLUG_ALREADY_EXISTS",
              "ISBN_ALREADY_EXISTS",
              "RESOURCE_IN_USE",
              "CONCURRENT_MODIFICATION",
              "CATEGORY_CYCLE",
              "BOOK_NOT_PUBLISHABLE",
              "STOCK_BELOW_RESERVED",
              "INSUFFICIENT_STOCK",
              "BOOK_NOT_AVAILABLE",
              "RESERVATION_MISMATCH",
              "RESERVATION_RELEASED",
              "RESERVATION_COMMITTED",
              "AUTHENTICATION_UNAVAILABLE",
              "INTERNAL_ERROR"
            ],
            "example": "SLUG_ALREADY_EXISTS"
          },
          "detail": {
            "type": "string",
            "description": "Genel açıklama; kullanıcı verisi içermez.",
            "example": "Slug is already in use."
          },
          "errors": {
            "type": "array",
            "description": "Yalnızca `VALIDATION_FAILED`'da: alan bazında hatalar. Gönderilen değer yer almaz.",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          },
          "instance": {
            "type": "string",
            "format": "uri-reference",
            "description": "İsteğin yolu.",
            "example": "/api/admin/publishers"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "HTTP durum kodu.",
            "example": 409
          },
          "title": {
            "type": "string",
            "description": "HTTP durumunun kısa adı.",
            "example": "Conflict"
          },
          "type": {
            "type": "string",
            "format": "uri",
            "description": "Problem tipi; varsayılan `about:blank` olduğunda yanıtta yer almaz."
          }
        },
        "required": [
          "code",
          "instance",
          "status",
          "title"
        ]
      },
      "PublisherRef": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        }
      },
...
      "UpdateBookRequest": {
        "type": "object",
        "properties": {
          "authorIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "maxItems": 20,
            "minItems": 0,
            "uniqueItems": true
          },
          "categoryIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "maxItems": 20,
            "minItems": 0,
            "uniqueItems": true
          },
          "coverUrl": {
            "type": "string",
            "maxLength": 500,
            "minLength": 0
          },
          "description": {
            "type": "string",
            "maxLength": 10000,
            "minLength": 0
          },
          "isbn": {
            "type": "string"
          },
          "pageCount": {
            "type": "integer",
            "format": "int32",
            "exclusiveMinimum": 0
          },
          "priceAmount": {
            "type": "number",
            "minimum": 0
          },
          "publisherId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "maxLength": 300,
            "minLength": 0,
            "pattern": ".*\\S.*"
          },
          "version": {
            "type": "integer",
            "format": "int64",
            "description": "Son okunan `version`; güncel değilse 409 `CONCURRENT_MODIFICATION`"
          }
        },
        "required": [
          "version"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "user-service `POST /api/auth/login` yanıtındaki `accessToken`; `role` = `ADMIN` olmalı.",
        "scheme": "bearer",
        "type": "http"
      },
      "internalApiKey": {
        "description": "Servisler arası ham anahtar (yalnızca çağıran serviste saklanır; catalog yalnızca SHA-256 özetini bilir). Kullanıcı JWT'si bu uçlarda geçersizdir.",
        "in": "header",
        "name": "X-Internal-Api-Key",
        "type": "apiKey"
      }
    }
  }
}